FilePath: Remove much of ToWStringHack, adding a LossyDisplayName()
The reason we don't want a free conversion between FilePaths and Unicode
is that it can be lossy.  But when displaying a string to the user,
we're ok if it's lossy when we have no other option.

This change introduces a LossyDisplayName() method that returns a
string16, and converts many of the users of ToWStringHack to use it.

BUG=69467

Review URL: http://codereview.chromium.org/6246036

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8a205c0f534bf8e5f17a531182dee38aaacd46e6