Add support for ICU data in a data file.
- Mac Foundation utility for fetching a resource from the MainAppBundle.
- Bringing in newer ICU dep that has optional support for using the ICU data
  file.
- If Mac is built using a ICU data file, directly map in the data so it will
  work even in a sandboxed process.

**The build does not use this data file support at this time.**

This might help the Mac bots as it means each unittest that depends on base
will be ~10meg smaller, helping link and also helping with the copy of bits
from the builder to the tester.

If Mac ever support a Universal binary (32bit and 64bit) it will also probably
want to use this to avoid carrying two copies of the ICU data.

Review URL: http://codereview.chromium.org/6479021

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 62e7190b0445dcc6f80b63149a8763f59c3a7cf0