Field trials are currently implemented (commonly) using a static variable that is set once, the first time it is necessary to decide if there is an experiment by a given name active.
With this change the field-test system can "push" a group that is selected for 
the given field trial (field test) if/when an experiment does arrive.

This change implements a simple IPC notification of the result of a FieldTrial
setting being sent to any previously started renderers.

BUG=16494
TEST=field trial tests
R=jar
Review URL: http://codereview.chromium.org/6883029

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 63a8ba1e1fb71156beff23b4a26828dbc387c734