Revert of 83710 which reverted 83641 - DNS Host resolver changes with retry logic. Fix for bug Chromium cannot recover from a state when its DNS requests have been dropped.
Whenever we try to resolve the host, we post a delayed task
to check if host resolution (OnLookupComplete) is completed 
or not. If it hasn't completed, then we start a new job 
to resolve for the same request.

BUG=73327
TEST=dns host resolver tests
R=eroman,jar
Review URL: http://codereview.chromium.org/6976006

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 189163ee560220770ed5a119f6cfb9244c50c39d