Added code to verify whether TerminateProcess is hooked before calling it.
This is to diagnose the bug referenced below. I think it is possible that a third party DLL is hooking TerminateProcess with a replacement that uses the wrong calling convention or has the wrong number of arguments, which causes Process::Terminate to return to the wrong address, causing a crash.

I put it in ThreadFunc because this is visible on the stack when the crash happens and because it is a windows specific .cc file.

BUG=81449

Review URL: http://codereview.chromium.org/7606016

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ef06bb6063e6b5b0f041ebbb8e28b1e51d2b02c4