Cleanup code used to diagnose and fix bug 81449.
I think it is unlikely that CloseHandle was the function that was being hooked because it is called from so many other places without problems. I am assuming that TerminateProcess was being hooked. If the symptoms of 81449 return when this is landed, this patch can be reverted.

BUG=81449

Review URL: http://codereview.chromium.org/7624052

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c4ed802942d46ce11a6d8728bb7b02110e6076a8