Profiling for the renderer process.
This lets you use --profiling-at-start=renderer and get profiler output for
each renderer process started on Linux.

The profiling classes are initialized at static initialization time, and the main thread is registered and the profile timer is started at that time. However, static initialization is done in the zygote, and the timer doesn't propagate across the fork, so trying to use the CPU profiler on the renderer never sampled. This patch re-registers the (new) main thread after the fork, which causes the profiler to check the state of the timer, and ends up re-starting it. A separate timer for flushing the profile and termination signal handler are added before the renderer process enters the sandbox in order to make sure the samples are written to disk.

BUG=none
TEST=Use it, get lovely traces, pperf away.


Review URL: http://codereview.chromium.org/8083025

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0a4cd3111631b21bdb07732ea037bbe40785590c