WriteInto handles length_with_null values of 1 better.
In debug builds on VS2010, the old code triggered a debug assertion in
xstring, since [0] was being accessed and the size of the array was 0.

A |length_with_null| of 1 is considered a valid use of this function, so
an alternate implementation is used which returns the data() pointer.

BUG=None
TEST=net_unittests --gtest_filter=*Digest* on debug on VS2010 works.


Review URL: http://codereview.chromium.org/8143001

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 11fe41a6b2cb4229f8ef3ff0bc22fd609fe838e4